Logica faltante backend

+++ src/main/java/com/greencycle/ecoswap/ecoswap/repository/OrdenRepository.java
@@ -4,7 +4,7 @@ import com.greencycle.ecoswap.ecoswap.model.Orden;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface OrdenRepository extends JpaRepository<Orden, Long> {
+public interface OrdenRepository extends JpaRepository<Orden, Integer> {
     // Para que la recicladora vea su historial
-    List<Orden> findByUsuarioId(Long usuarioId);
+    List<Orden> findByUsuarioId(Integer usuarioId);
 }
